+++ CHANGELOG.md
@@ -1,15 +1,20 @@
 # Change Log
 
+
 ## [3.0.4] - SNAPSHOT
 
 ### Added
 - Aggiunto store type SAVE\_ORPHAN\_PEC\_RECEIPT\_AS\_VARIE per il salvataggio di ricevute PEC orfane come documenti non protocollati (richiede la property '_docway4mailboxmanager.pec.orphan-receipts-as-varie_')
+
+### Changed
+- Modificata la logica dei contatori su messaggi elaborati, skippati e in errore processati su una specifica mailbox
 
 ### Fixed
 - In generazione documenti da fatturePA, compilazione dei campi del rif. esterno con i dati estratti dalla fattura stessa (e non recuperati da ACL) (Task #16546)
 - Corretto indice del messaggio in elaborazione sul log di MSA
 - Corretta la registrazione dell'indirizzo mail del mittente su audit di MSA
 - Corretto i parametri di configurazione per il protocollo SMTP-TLS (nuova versione mail.jar)
+
 
 ## [3.0.3] - 2019-01-07
 
@@ -25,10 +30,12 @@
 - Corretto bug in individuazione di files definiti su segnatura.xml (validazione della segnatura)
 - Corretti xpath di estrazione Identificatore da notifiche relative ad interoperabilità
 
+
 ## [3.0.2] - 2018-12-07 
 
 ### Fixed
 - Aggiornata la versione della libreria mail
+
 
 ## [3.0.1] - 2018-10-17 
 
@@ -39,6 +46,7 @@
 
 ### Fixed
 - Corretto comportamento dello store type UPDATE\_NEW\_RECIPIENT: gli RPA delle caselle successive (alla prima) vengono riportati come CC con diritto di intervento
+
 
 ## [3.0.0] - 2018-09-06 
 
@@ -55,6 +63,7 @@ alle caselle di posta
 - La configurazione della protocollazione delle fatturePA deve essere recuperata dalla configurazione della casella di posta su eXtraWay e non dal file di
 properties di MSA
 
+
 ## [1.0.2] - 2018-06-27
 
 ### Changed
@@ -62,6 +71,7 @@ properties di MSA
 
 ### Fixed
 - Corretto problema di estrazione del mittente da ACL nel caso di PEC (l'indirizzo del mittente per la ricerca in ACL ora viene estratto da daticert.xml)
+
 
 ## [1.0.1] - 2018-05-25 
 
@@ -71,6 +81,7 @@ properties di MSA
     - docway4mailboxmanager.mail-sender.socket-timeout=60000
     - docway4mailboxmanager.mail-sender.connection-timeout=10000
 - In DocwayMailboxManager realizzato codice per eseguire n-tentativi in fase di connessione al server smtp
+
 
 ## [1.0.0] - 2018-05-13